+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v 1.0.0 (?)
 Changes in this release:
 - Added error highlighting (#18)
+- Added export to server button (#8)
 - Use separate virtual env for the compiler (#9)
 
 # v 0.0.1